backend/requirements.txt
Flask==2.1.2
Flask-Cors==3.0.10
pandas
pandas-ta
numpy
matplotlib
seaborn
python-binance
tabulate
python-dotenv
